Exit with non-zero exit code when compilation fails (#120)
At the moment, when compilation fails we get an error message back
with an exception name in it but a non-zero exit code. This changes it
to have a similar error message but exits with a non-zero exit code.

This probably isn't the right long-term solution but simplifies the code
slightly and fixes a bug in the meantime.